functions/sftp-operations/requirements.txt
azure-functions
azure-identity
azure-keyvault-secrets
azure-storage-blob
azure-servicebus
paramiko
cryptography
pynacl
python-gnupg
pgpy